requirements.txt
Flask==2.0.3
flask_bcrypt==1.0.1
flask_login==0.6.0
flask_mail==0.9.1
Flask_SQLAlchemy==2.5.1
flask_wtf==1.0.1
itsdangerous==2.1.0
Pillow==9.1.0
WTForms==3.0.1
gunicorn==20.1.0
